+++ app/src/main/java/com/moo/mool/view/comment/model/RequestPutEmoji.kt
@@ -1,4 +1,4 @@
-package com.moo.mool.view.comment
+package com.moo.mool.view.comment.model
 
 import com.google.gson.annotations.SerializedName
