+++ app/src/main/java/com/moo/mool/view/vote/model/RequestPostComment.kt
@@ -1,4 +1,4 @@
-package com.moo.mool.view.vote
+package com.moo.mool.view.vote.model
 
 data class RequestPostComment(
     val commentLayer: Int = 0,
