+++ app/src/main/java/com/moo/mool/view/vote/viewmodel/VoteViewModel.kt
@@ -3,12 +3,12 @@ package com.moo.mool.view.vote.viewmodel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import com.moo.mool.view.comment.Comment
-import com.moo.mool.view.comment.RequestPutEmoji
-import com.moo.mool.view.vote.RequestCommentId
-import com.moo.mool.view.vote.RequestPostComment
-import com.moo.mool.view.vote.RequestVote
-import com.moo.mool.view.vote.ResponseFeedDetail
+import com.moo.mool.view.comment.model.Comment
+import com.moo.mool.view.comment.model.RequestPutEmoji
+import com.moo.mool.view.vote.model.RequestCommentId
+import com.moo.mool.view.vote.model.RequestPostComment
+import com.moo.mool.view.vote.model.RequestVote
+import com.moo.mool.view.vote.model.ResponseFeedDetail
 import com.moo.mool.view.vote.repository.VoteRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
