+++ app/src/main/res/layout/item_comment.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="comment"
-            type="com.moo.mool.view.comment.Comment" />
+            type="com.moo.mool.view.comment.model.Comment" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
